refactor capture utils; defensive payload size limit check in post-decode b64 mirror codepath
